Get things working for the demo
I added a new NF called slow_speed_tester. Currently it doesn't do
anything different form speed_tester (lol). The gist of the demo
is for slow_speed_tester to send a packet to the fork_tester, which
does more complex calculations each time it receives a packet. FT
sends the packet back to SST, creating a loop. Once the CPU usage
of FT goes above a thereshold, it forks to a new process.

I want to make it so that we wait for the forked process to be ready,
but currently we just sleep for a few seconds once the fork is made
in order to let it set up.